account-vault-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
pydantic==2.8.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
cryptography==43.0.1
python-dotenv==1.0.1
psycopg[binary]==3.2.9
